Safari-611.1.10/LayoutTests/imported/w3c/web-platform-tests/resource-timing/TAO-case-insensitive-null-opaque-origin.sub-expected.txt

Description

This test validates that for a cross origin resource, the timing allow check algorithm will fail when the value of Timing-Allow-Origin is a case-insensitive match to null and the origin is an opaque origin.


FAIL Test case-insensitive null TAO value with opaque origins assert_equals: expected "PASS" but got "FAIL"
